Update to Zero 0.6 - 3/4
The rest of zero.config.ts goes away. Zero no longer has a config
file. Configuration is done either via CLI flags or environment
variables.

This commit moves the config values into the checked-in .env file.

Note that the names are now standardized - zero-cache itself looks
for these env vars, rather than the app dev being able to map
whatever env vars it wants into the config like before.

Also as part of this change, a few cleanups:

* Update all the places that referred to these env vars.

* The old sample used the `postgres` (master) pg db for upstream
data, as well as zero cvr/db data. This was a bit odd and
unrealistic. This change uses a separate pg database (but not
server/cluster) for each of these needs.

* The old sample used port 6543 for the db on the docker host side
but this required port mapping and an extra env var. Instead,
just use the default 5432 port and get rid of the mapping.